Clean frontend tests. Fix build issues
