--- src/Frontend/test/specs/authentication/auth-invalid-redirect.spec.ts
@@ -1,46 +1,0 @@
-import { vi, expect } from "vitest";
-import { createOidcMockWithInvalidRedirectUri } from "../../mocks/oidc-client-mock";
-
-// Mock oidc-client-ts with signinCallback that fails (invalid redirect URI)
-vi.mock("oidc-client-ts", () => createOidcMockWithInvalidRedirectUri());
-
-import { test, describe } from "../../drivers/vitest/driver";
-import * as precondition from "../../preconditions";
-import { waitFor } from "@testing-library/vue";
-
-describe("FEATURE: Invalid Redirect URI (Scenario 11)", () => {
-  describe("RULE: App should handle OAuth callback errors gracefully", () => {
-    test("EXAMPLE: Failed signinCallback prevents authentication", async ({ driver }) => {
-      await driver.setUp(precondition.serviceControlWithMonitoring);
-      await driver.setUp(precondition.hasAuthenticationEnabled());
-
-      // Navigate to trigger auth flow
-      // Since getUser returns null and signinCallback fails,
-      // user should not be authenticated
-      await driver.goTo("/dashboard");
-
-      // With no valid user session, auth token should not be set
-      await waitFor(() => {
-        expect(sessionStorage.getItem("auth_token")).toBeNull();
-      });
-    });
-
-    test("EXAMPLE: signinRedirect is called when user has no session", async ({ driver }) => {
-      await driver.setUp(precondition.serviceControlWithMonitoring);
-      await driver.setUp(precondition.hasAuthenticationEnabled());
-
-      // Navigate to protected route without authentication
-      await driver.goTo("/dashboard");
-
-      // Since getUser returns null, the app should redirect to IdP
-      // Auth token should not be set
-      await waitFor(() => {
-        expect(sessionStorage.getItem("auth_token")).toBeNull();
-      });
-
-      // Note: In a real scenario with invalid redirect URI configured,
-      // the IdP would reject the redirect and display its own error.
-      // This test verifies the app correctly initiates the redirect flow.
-    });
-  });
-});
