+++ src/Frontend/test/specs/authentication/auth-session-storage.spec.ts
@@ -8,7 +8,28 @@ import { test, describe } from "../../drivers/vitest/driver";
 import * as precondition from "../../preconditions";
 import { waitFor, screen } from "@testing-library/vue";
 
-describe("FEATURE: Session Persistence (Scenario 6)", () => {
+describe("FEATURE: Session Storage Behavior", () => {
+  describe("RULE: Sessions should be isolated per browser tab", () => {
+    test("EXAMPLE: Auth token is stored in sessionStorage (tab-specific, not shared)", async ({ driver }) => {
+      await driver.setUp(precondition.serviceControlWithMonitoring);
+      await driver.setUp(precondition.hasAuthenticationEnabled());
+
+      await driver.goTo("/dashboard");
+
+      await waitFor(() => {
+        expect(screen.getByText(/Dashboard/i)).toBeInTheDocument();
+      });
+
+      // Verify token is in sessionStorage (tab-specific)
+      const sessionToken = sessionStorage.getItem("auth_token");
+      expect(sessionToken).toBe(defaultMockUser.access_token);
+
+      // Verify token is NOT in localStorage (would be shared across tabs)
+      const localToken = localStorage.getItem("auth_token");
+      expect(localToken).toBeNull();
+    });
+  });
+
   describe("RULE: Session should persist across navigation within the same tab", () => {
     test("EXAMPLE: Auth token persists when navigating between pages", async ({ driver }) => {
       await driver.setUp(precondition.serviceControlWithMonitoring);
